Frisbee/requirements.txt
Django==1.9.2
django-grappelli==2.9.1
django-nested-admin==2.1.12
django-suit==0.2.15
django-super-inlines==0.1.4
django-tables2==1.1.1
django-autocomplete-light


